move code to ramn_j1939.c
