+++ firmware/RAMNV1/Core/Inc/ramn_config.h
@@ -194,7 +194,7 @@
 
 // Enable this flag to force autopilot/fuzzer mode.
 // When enabled, ECU A will periodically randomize command values at 1 Hz
-// and other ECUs will follow those commands received via CAN, INSTEAD OF thei
+// and other ECUs will follow those commands received via CAN, INSTEAD OF their
 // sensor inputs. This relies on the CARLA driving commands but
 // without requiring UDS activation or physical sensor inputs.
 // #define RAMN_FORCE_AUTOPILOT
