+++ firmware/RAMNV1/Core/Inc/ramn_j1939.h
@@ -1,6 +1,8 @@
 #ifndef INC_RAMN_J1939_H_
 #define INC_RAMN_J1939_H_
 
+#include "main.h"
+
 /* J1939 PGN Definitions */
 #define J1939_PGN_TSC1              0       /* Torque/Speed Control 1 */
 #define J1939_PGN_TC1               256     /* Transmission Control 1 */
@@ -109,5 +111,9 @@
 /* Proprietary A Payload Definitions */
 /* Command_Steering (PGN 61184): SA AEBS (160), DA Steering Controller (19). Bytes 1-2. j1939_val = ramn_val */
 /* Command_Horn (PGN 61184): SA Shift Console (5), DA Body Controller (33). Byte 1. j1939_val = ramn_val & 0xFF */
+
+#ifdef ENABLE_J1939_MODE
+void RAMN_J1939_ProcessRxCANMessage(const FDCAN_RxHeaderTypeDef* pHeader, const uint8_t* data);
+#endif
 
 #endif /* INC_RAMN_J1939_H_ */
